refactor(headless): extract fixed prompt WAL types
Move fixed-prompt WAL schema constants and event types into a package-local leaf, then migrate production schema consumers to depend on it directly.

Keep controller imports compatible through explicit re-exports and add a source contract for ownership and dependency direction. The JSONL schema and runtime behavior are unchanged.
